requirements.txt
selenium
pandas
openpyxl
numpy
jpype1
requests
tabula-py